Fix time conversion from UTC to localtime
When changing from summer to winter time, the conversion form UTC
to locatime also changes (i.e. UTC+2 to UTC+1). This conversion is now
implemented using the pandas functions dt_localize and dt_convert to
automatically convert form UTC to the local time.
Note that the database timestamp is always in UTC!